update content in main.rb
