--- lib/board.rb
@@ -1,11 +1,0 @@
-class Board
-  board = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-
-  puts '+---+---+---+'
-  puts "| #{board[0]} | #{board[1]} | #{board[2]} |"
-  puts '+---+---+---+'
-  puts "| #{board[3]} | #{board[4]} | #{board[5]} |"
-  puts '+---+---+---+'
-  puts "| #{board[6]} | #{board[7]} | #{board[8]} |"
-  puts '+---+---+---+'
-end
