--- lib/player.rb
@@ -1,8 +1,0 @@
-class Player
-  attr_accessor :player1, :player2
-
-  def initialize
-    @player1 = :player1
-    @player2 = :player2
-  end
-end
